Chore: "reservation2 -> reservation으로 변경"
- 기존의 reservation 패키지 삭제 후 reservation2 패키지와 하위 컴포넌트들 이름을 reservation으로 변경
