+++ hotelity/src/component/hotel-service/reservation/Reservation.vue
@@ -1,0 +1,16 @@
+<template>
+  <div>
+    <ReservationSidebar/>
+    <ReservationNavbar/>
+  </div>
+
+</template>
+
+<script setup>
+import ReservationSidebar from "@/component/hotel-service/reservation/ReservationSidebar.vue";
+import ReservationNavbar from "@/component/hotel-service/reservation/ReservationNavbar.vue";
+</script>
+
+<style>
+
+</style>
