+++ hotelity/src/component/hotel-service/reservation/ReservationCalendar.vue
@@ -2,72 +2,130 @@
   <div id='calendar'></div>
 </template>
 
-<script>
+<script setup>
 
-  import { defineComponent } from 'vue';
-  import { Calendar } from '@fullcalendar/core';
-  import dayGridPlugin from '@fullcalendar/daygrid';
-  import interactionPlugin from '@fullcalendar/interaction';
+import {ref, onMounted, watch, inject} from 'vue';
+import {Calendar} from '@fullcalendar/core';
+import dayGridPlugin from '@fullcalendar/daygrid';
+import interactionPlugin from '@fullcalendar/interaction';
 
-  let selectedDay = null;
+const emit = defineEmits();
+const props = defineProps({
+  reservations: Object
+});
 
-  export default defineComponent({
-    setup(_, { emit }) { // Add the emit function here
-      const dateClick = async function(info) {
-        if(selectedDay) {
-          selectedDay.style.borderColor = '';
-          selectedDay.style.borderWidth = '';
-        }
-        // 선택 시 css
-        info.dayEl.style.borderColor = 'skyblue';
-        info.dayEl.style.borderWidth = '5px';
-        selectedDay = info.dayEl;
+var selectedMonth = ref(null);
+var selectedDay = ref(null);
 
-        // 선택 시 메소드 호출
-        const formattedDate = formatDateTime(new Date(info.dateStr));
-
-        console.log("변환된 날짜");
-        console.log(formattedDate)
-
-        emit ('date-Clicked', formattedDate);
-      };
-
-      return {
-        dateClick,
-      };
-    },
-    mounted() {
-      const calendarEl = document.getElementById('calendar'); // FullCalendar를 삽입할 요소를 가져옵니다.
-      let calendar = new Calendar(calendarEl, {
-        plugins: [dayGridPlugin, interactionPlugin],
-        initialView: 'dayGridMonth',
-        selectable: true,
-        events: [
-          {
-            title: '예약 n건',
-            start: '2024-06-03'
-          },
-          {
-            title: '예약 n건',
-            start: '2024-06-04'
-          }
-        ],
-        dateClick: this.dateClick,
-        aspectRatio: 3.5
-      });
-      calendar.render(); // FullCalendar를 렌더링합니다.
-    },
-  });
-
-// Datepicker 값 -> yyyy-MM-ddTHH:mm:ss(LocalDateTime)로 변환
-  function formatDateTime(date) {
-    const year = date.getFullYear();
-    const month = (date.getMonth() + 1).toString().padStart(2, '0');
-    const day = date.getDate().toString().padStart(2, '0');
-    const hours = date.getHours().toString().padStart(2, '0');
-    const minutes = date.getMinutes().toString().padStart(2, '0');
-    const seconds = date.getSeconds().toString().padStart(2, '0');
-
-    return `${year}-${month}-${day}T${hours}:${minutes}:${seconds}`;
+// 캘린더 날짜 선택 시 스타일 적용
+const dateClick = async function (info) {
+  if (selectedDay.value) {
+    selectedDay.value.style.borderColor = '';
+    selectedDay.value.style.borderWidth = '';
   }
+  // 선택 시 css
+  info.dayEl.style.borderColor = 'skyblue';
+  info.dayEl.style.borderWidth = '5px';
+  selectedDay.value = info.dayEl;
+
+  // 선택 시 메소드 호출
+  const formattedDate = formatDate(new Date(info.dateStr));
+
+  // console.log("변환된 날짜");
+  // console.log(formattedDate)
+
+  emit('date-clicked', formattedDate);
+};
+
+// 예약 건 수 계산을 위한 변수
+let response = inject('response');
+
+// 부모 컴포넌트의 메소드 호출 감지 변수
+let change = inject('change');
+
+// 예약 건수 계산
+watch(response, (newResponse) => {
+  if (newResponse) {
+    const events = newResponse.data.data.content ? Object.entries(newResponse.data.data.content).map(([date, reservations]) => ({
+      title: `예약 ${reservations.length} 건`,
+      start: date
+    })) : [];
+
+    const eventSource = calendar.getEventSourceById('reservations');
+    if (eventSource) {
+      eventSource.remove();
+    }
+    calendar.addEventSource({events, id: 'reservations'});
+  }
+});
+
+let calendar = null;
+
+// 캘린더 이벤트 업데이트
+const updateEvents = () => {
+  console.log("updateEvents 실행됨")
+  const events = props.reservations ? Object.entries(props.reservations).map(([date, reservations]) => ({
+    title: `예약 ${reservations.length} 건`,
+    start: date
+  })) : [];
+
+  const eventSource = calendar.getEventSourceById('reservations');
+  if (eventSource) {
+    eventSource.remove();
+  }
+  console.log("addEventSource 실행됨")
+  calendar.addEventSource({events, id: 'reservations'});
+};
+
+// 페이지 열리면 캘린더 렌더링
+const renderCalendar = () => {
+  const calendarEl = document.getElementById('calendar');
+  calendar = new Calendar(calendarEl, {
+    plugins: [dayGridPlugin, interactionPlugin],
+    initialView: 'dayGridMonth',
+    selectable: true,
+    events: [],
+    dateClick: dateClick,
+    aspectRatio: 3.5,
+    datesSet: function (info) {
+
+      let firstWeekStartDate = new Date(info.startStr);
+      firstWeekStartDate.setDate(firstWeekStartDate.getDate() + 7);
+      let year = firstWeekStartDate.getFullYear();
+      let month = String(firstWeekStartDate.getMonth() + 1).padStart(2, '0'); // Months are 0-based, so we add 1
+      let day = String(firstWeekStartDate.getDate()).padStart(2, '0');
+      selectedMonth.value = `${year}-${month}-${day}T00:00:00`;
+      emit('month-changed', selectedMonth.value);
+    }
+  });
+  calendar.render();
+  updateEvents();
+};
+
+// 부모 컴포넌트(ReservationList.vue) 백엔드 메소드 호출 감지
+watch(change, (newValue) => {
+  if (newValue) {
+    if (!calendar) {
+      renderCalendar();
+    } else {
+      updateEvents();
+    }
+  }
+});
+
+// 페이지 접속 시점에 캘린더 렌더링
+onMounted(renderCalendar);
+
+// 날짜 형식 변환(FullCalendar의 date형식 -> yyyy-MM-dd)
+function formatDate(date) {
+  if (date == null) {
+    date = new Date();
+  }
+
+  const year = date.getFullYear();
+  const month = String(date.getMonth() + 1).padStart(2, '0'); // Months are 0-based
+  const day = String(date.getDate()).padStart(2, '0');
+  return `${year}-${month}-${day}`;
+}
+
 </script>
