+++ hotelity/src/component/hotel-service/reservation/ReservationList.vue
@@ -1,6 +1,6 @@
 <template>
   <div class="container-fluid pt-4 px-4">
-    <ReservationCalendar @date-clicked="handleDateClicked"/>
+    <Reservation2Calendar :reservations="obj" @month-changed="updateSelectedMonth" @date-clicked="handleDateClicked"/>
   </div>
 
   <!-- Table Start -->
@@ -8,7 +8,7 @@
     <div class="bg-secondary rounded-top p-4">
       <h3 class="mb-4">예약 리스트</h3>
       <div class="search-container d-flex align-items-center">
-        <ReservationSearch/>
+        <Reservation2Search/>
       </div>
       <div class="position-relative-container mt-3">
         <ExcelButton/>
@@ -18,10 +18,11 @@
           <button class="btn btn-secondary" style="background-color: saddlebrown;" @click="toggleCalendarContainer"><i
               class="bi bi-calendar"></i></button>
 
-          <ReservationCheckinBtn :checkedRows="checkedRows" :reservations="reservations"/>
+          <Reservation2CheckinBtn :checkedRows="checkedRows" :reservations="reservations"/>
 
           <!--        ReservationFilter start -->
-          <button id="filter-icon" class="btn btn-secondary" style="background-color: saddlebrown;" @click="toggleFilterContainer"><i
+          <button id="filter-icon" class="btn btn-secondary" style="background-color: saddlebrown;"
+                  @click="toggleFilterContainer"><i
               class="bi bi-funnel"></i></button>
         </div>
         <div class="filter-container" style="width: auto" v-show="isFilterContainerVisible">
@@ -73,7 +74,7 @@
               <th scope="col">객실 코드</th>
               <th scope="col">객실명</th>
               <th scope="col">객실 등급</th>
-              <!--              <th scope="col">객실 수용 인원</th>-->
+              <!--<th scope="col">객실 수용 인원</th>-->
               <th scope="col">지점 코드</th>
               <th scope="col">예약 일자</th>
               <th scope="col">예약 체크인 일자</th>
@@ -84,10 +85,10 @@
             </tr>
             </thead>
             <tbody>
-            <tr v-for="(reservation, index) in reservations.content" :key="reservation.reservationCodePk">
+            <tr v-for="(reservation, index) in reservations" :key="reservation.reservationCodePk">
               <td>
                 <input type="checkbox" :checked="reservation.stayStatus === 1" :disabled="reservation.stayStatus === 1"
-                       @change="checkedRows[index] = $event.target.checked ? reservation.reservationCodePk : null">
+                       @change="checkedRows[index] = $event.target.checked ? reservation : null">
               </td>
               <td>{{ reservation.reservationCodePk }}</td>
               <td>{{ reservation.customerCodeFk }}</td>
@@ -96,7 +97,6 @@
               <td>{{ reservation.roomCodeFk }}</td>
               <td>{{ reservation.roomName }}</td>
               <td>{{ reservation.roomLevelName }}</td>
-              <!--              <td>{{ reservation.roomCapacity }}</td>-->
               <td>{{ reservation.branchCodeFk }}</td>
               <td>{{ formatDate(reservation.reservationDate) }}</td>
               <td>{{ formatDate(reservation.reservationCheckinDate) }}</td>
@@ -121,17 +121,13 @@
 
 <script setup>
 
-import {ref, defineComponent, onMounted, watch} from 'vue';
-import axios from 'axios';
+import {ref, onMounted, watch, provide} from 'vue';
+import * as api from '@/api/apiService.js';
 
-import ReservationSearch from "@/component/hotel-service/reservation/ReservationSearch.vue";
 import ExcelButton from "@/component/common/ExcelButton.vue";
-import ReservationFilter from "@/component/hotel-service/reservation/ReservationFilter.vue";
 import DatePicker from "vue3-datepicker";
-import ReservationCheckinBtn from "@/component/hotel-service/reservation/ReservationCheckinBtn.vue";
-import StayCheckoutBtn from "@/component/hotel-service/stay/StayCheckoutBtn.vue";
-import ReservationCalendar from "@/component/hotel-service/reservation/ReservationCalendar.vue";
-
+import Reservation2Calendar from "@/component/hotel-service/reservation/ReservationCalendar.vue";
+import Reservation2CheckinBtn from "@/component/hotel-service/reservation/ReservationCheckinBtn.vue";
 
 const isLoading = ref(true);
 const reservations = ref([]);
@@ -140,75 +136,70 @@ const selectedReservationCheckinDate = ref(null);
 const selectedReservationCheckoutDate = ref(null);
 const isFilterContainerVisible = ref(false);
 const isCalendarContainerVisible = ref(false);
+const selectedMonth = ref(null);
 
 // 체크박스
 const checkedRows = ref([]);
 
+// reservations 변경 감지
 watch(reservations, () => {
-  checkedRows.value = reservations.value.content ?
-      reservations.value.content.map(reservation =>
-          reservation.stayStatus === 1 ?
-              reservation.reservationCodePk : null) : [];
+  if (reservations.value && reservations.value.content) {
+    checkedRows.value = reservations.value.content.map(reservation =>
+        reservation.stayStatus === 1 ?
+            reservation.reservationCodePk : null);
+  }
 }, {immediate: true});
 
-// 월별 예약 내역 조회
-async function fetchData(params) {
-  try {
-    const formattedDate = formatDateTime(selectedReservationCheckinDate.value);
-    const response = await axios.get(`http://localhost:8888/hotel-service/reservations/${formattedDate}`, {params});
-    console.log("월별 예약 내역 조회: ")
-    console.log(response.data);
-    return response.data.data;
-  } catch (error) {
-    console.error(error);
+
+// 캘린더 월 변경 감지
+function updateSelectedMonth(newMonth) {
+  selectedMonth.value = newMonth;
+}
+
+// 캘린더 날짜 선택 감지
+function handleDateClicked(formattedDate) {
+  if (obj[formattedDate]) {
+    console.log(obj[formattedDate]);
+    reservations.value = obj[formattedDate];
+  } else {
+    console.log("No reservations for " + formattedDate);
+    reservations.value = [];
   }
 }
 
-// 일별 예약 내역 조회
-const handleDateClicked = async (formattedDate) => {
-  console.log("handleDateClicked 실행됨")
-  try {
-    const response = await axios.get(`http://localhost:8888/hotel-service/reservations/${formattedDate}/day`);
-    console.log("일별 예약 내역 조회: ")
-    console.log(response.data.data);
-    reservations.value = response.data.data;
-  } catch (error) {
-    console.error(error);
-  }
-}
+let obj = {};
+const change = ref(false);
+provide('change', change);
 
-watch(selectedReservationCheckinDate, loadReservations, {immediate: true});
+watch(selectedMonth, async (newMonth) => {
+      if (newMonth) {
+        const response = await api.getMonthlyReservations(selectedMonth.value);
+        const data = response.data;
 
-async function loadReservations() {
-  console.log('selectedReservationCheckinDate: ', selectedReservationCheckinDate.value);
-  reservations.value = await fetchData({
-    reservationCodePk: null,
-    customerCodeFk: null,
-    customerName: null,
-    customerEnglishName: null,
-    roomCodeFk: null,
-    roomName: null,
-    roomLevelName: null,
-    roomCapacity: null,
-    branchCodeFk: null,
-    reservationDate: null,
-    reservationCheckinDate: null,
-    reservationCheckoutDate: null,
-    reservationCancelStatus: null,
-    reservationPersonnel: null,
-    stayStatus: null,
-    orderBy: null,
-    sortBy: null
-  });
-  console.log("확인 요망")
-  console.log(reservations.value);
+        change.value = true;
 
-  isLoading.value = false;
-}
+        console.log("백엔드 메소드 실행 결과: ")
+        console.log(data);
+
+        for (const reservation of data.content) {
+          const date = formatDate(reservation.reservationCheckinDate);
+
+          if (!obj[date]) {
+            obj[date] = [];
+          }
+
+          obj[date].push(reservation);
+        }
+
+        console.log(obj);
+      }
+    }
+);
 
 // 필터
 function toggleFilterContainer() {
   isFilterContainerVisible.value = !isFilterContainerVisible.value;
+
   // 필터 컨테이너가 열리면 캘린더 컨테이너는 자동 닫힘
   if (isFilterContainerVisible.value && isCalendarContainerVisible.value) {
     isCalendarContainerVisible.value = false;
@@ -217,7 +208,7 @@ function toggleFilterContainer() {
 
 function toggleCalendarContainer() {
   isCalendarContainerVisible.value = !isCalendarContainerVisible.value;
-  if(isCalendarContainerVisible.value && isFilterContainerVisible.value) {
+  if (isCalendarContainerVisible.value && isFilterContainerVisible.value) {
     isFilterContainerVisible.value = false;
   }
 }
@@ -226,8 +217,19 @@ function toggleDropdownMenu() {
   isDropdownOpen.value = !isDropdownOpen.value;
 }
 
+function getToday() {
+  const today = new Date();
+  const year = today.getFullYear();
+  const month = String(today.getMonth() + 1).padStart(2, '0'); // Months are 0-based
+  const day = String(today.getDate()).padStart(2, '0');
+
+  return `${year}-${month}-${day}T00:00:00`;
+}
+
 onMounted(async () => {
-  // await loadReservations();
+  const today = getToday();
+  const response = await api.getDailyReservations(today);
+  reservations.value = response.data.content;
   $('#filter-icon').on('click', function () {
     $('#filter').toggle();
   });
@@ -262,8 +264,8 @@ function formatDateTime(date) {
 <style>
 .filter-container {
   position: absolute;
-  top: 50px;  /* 필터 아이콘의 높이에 따라 조정 */
-  right: -8px;  /* 필터 아이콘 오른쪽 끝에 위치 */
+  top: 50px; /* 필터 아이콘의 높이에 따라 조정 */
+  right: -8px; /* 필터 아이콘 오른쪽 끝에 위치 */
   width: auto;
 }
 
@@ -275,9 +277,8 @@ function formatDateTime(date) {
   background: #fff;
   border-radius: 5px;
   padding: 10px;
-  box-shadow: 0px 8px 16px 0px rgba(0,0,0,0.2); /* 그림자 효과 추가 */
+  box-shadow: 0px 8px 16px 0px rgba(0, 0, 0, 0.2); /* 그림자 효과 추가 */
 }
-
 
 .calendar-container::before {
   content: "";
