--- hotelity/src/component/hotel-service/reservation2/Reservation.vue
@@ -1,16 +1,0 @@
-<template>
-  <div>
-    <Reservation2Sidebar/>
-    <Reservation2Navbar/>
-  </div>
-
-</template>
-
-<script setup>
-import Reservation2Sidebar from "@/component/hotel-service/reservation2/Reservation2Sidebar.vue";
-import Reservation2Navbar from "@/component/hotel-service/reservation2/Reservation2Navbar.vue";
-</script>
-
-<style>
-
-</style>
