+++ hotelity/src/router/index.js
@@ -25,7 +25,7 @@ import TestCustomer from '@/component/TestCustomer.vue';
 import NoticeInfo from '@/component/sales/NoticeInfo.vue';
 import RoomDetails from '@/component/hotel-management/RoomDetails.vue';
 import Stay from "@/component/hotel-service/stay/Stay.vue";
-import Reservation from "@/component/hotel-service/reservation2/Reservation.vue";
+import Reservation from "@/component/hotel-service/reservation/Reservation.vue";
 
 const router = createRouter({
   history: createWebHistory(),
