+++ hotelity/src/router/router.js
@@ -19,8 +19,6 @@ import RoomDetails from "@/component/hotel-management/RoomDetails.vue";
 import FacilityList from "@/component/hotel-management/FacilityList.vue";
 import Ancillary from "@/component/hotel-management/Ancillary.vue";
 
-import ReservationPage from '@/component/hotel-service/reservation/ReservationPage.vue';
-import Reservation2List from '@/component/hotel-service/reservation2/Reservation2List.vue';
 import ReservationList from "@/component/hotel-service/reservation/ReservationList.vue";
 import StayList from "@/component/hotel-service/stay/StayList.vue";
 import StayPage from "@/component/hotel-service/stay/StayPage.vue";
@@ -172,15 +170,6 @@ const routes = [
         path: '/reservation',
         name: 'Reservation',
 
-        component: Reservation2List,
-        meta: {
-            requiresAuth: true,
-            category: 'hotelService',
-        },
-    },
-    {
-        path: '/reservationList',
-        name: 'ReservationList',
         component: ReservationList,
         meta: {
             requiresAuth: true,
